mvpscripts/requirements.txt
aws-cdk-lib==2.12.0
constructs>=10.0.0,<11.0.0
cdk-ec2-key-pair==3.2.0
cdk-iam-floyd>=0.300.0